+++ cmd/ctx/main.go
@@ -6,7 +6,6 @@ import (
 	"fmt"
 	"os"
 
-	"github.com/fatih/color"
 	"github.com/nochso/ctxerr"
 )
 
@@ -43,7 +42,7 @@ Flags:
 func main() {
 	context := flag.Int("context", ctxerr.DefaultContext, "print `NUM` lines of context surrounding an error. negative for all, positive for limited and 0 for none (default)")
 	pessimistic := flag.Bool("pessimistic", false, "print only matching errors, ignore everything else")
-	flag.BoolVar(&color.NoColor, "no-color", false, "disable any color output")
+	flag.BoolVar(&ctxerr.NoColor, "no-color", false, "disable any color output")
 	flag.Usage = usage
 	flag.Parse()